+++ prod-sample/sample-modules/frank-lloyd-root/0.0.0/src/components/FrankLloydRoot.jsx
@@ -21,7 +21,7 @@ import ImmutablePropTypes from 'react-immutable-proptypes';
 import { fromJS } from 'immutable';
 import ModuleRoute from 'holocron-module-route';
 import { holocronModule } from 'holocron';
-import Helmet from 'react-helmet';
+import { Helmet } from 'react-helmet';
 import { connect } from 'react-redux';
 import { compose } from 'redux';
 import HelloWorldComponent from './HelloWorld';
@@ -42,13 +42,13 @@ export function FrankLloydRoot({ children, config }) {
   );
 }
 
-FrankLloydRoot.childRoutes = (store) => ([
+FrankLloydRoot.childRoutes = (store) => [
   <ModuleRoute path="vitruvius" moduleName="vitruvius-franklin" />,
   <ModuleRoute path="success" component={HelloWorldComponent} />,
   <ModuleRoute path="healthy-frank" moduleName="healthy-frank" />,
   <ModuleRoute path="demo/:moduleName" moduleName="preview-frank" />,
   <ModuleRoute path="html-partial/:locale/:moduleName" component={Partial} onEnter={onPartialRouteEnter(store)} />,
-]);
+];
 
 FrankLloydRoot.propTypes = {
   children: PropTypes.node.isRequired,