+++ prod-sample/sample-modules/frank-lloyd-root/0.0.3/src/components/FrankLloydRoot.jsx
@@ -20,12 +20,12 @@ import PropTypes from 'prop-types';
 import ImmutablePropTypes from 'react-immutable-proptypes';
 import { fromJS } from 'immutable';
 import ModuleRoute from 'holocron-module-route';
-import Helmet from 'react-helmet';
+import { Helmet } from 'react-helmet';
 import { connect } from 'react-redux';
 import HelloWorldComponent from './HelloWorld';
 
 export function FrankLloydRoot({ children, config }) {
-  const cdnUrl = config.get('cdnUrl');
+  const cdnUrl = React.useMemo(() => config.get('cdnUrl'), [config]);
   return (
     <React.Fragment>
       <Helmet
@@ -36,7 +36,6 @@ export function FrankLloydRoot({ children, config }) {
           { name: 'theme-color', content: '#FDB92D' },
         ]}
         link={[
-          { rel: 'webmanifest', href: `${cdnUrl}modules/frank-lloyd-root/0.0.3/assets/manifest.webmanifest` },
           // favicon
           { rel: 'icon', href: `${cdnUrl}modules/frank-lloyd-root/0.0.3/assets/pwa-icon-192px.png` },
           // icon ideally with a size of 192px (or 180px), it is added as the home icon
@@ -49,12 +48,12 @@ export function FrankLloydRoot({ children, config }) {
   );
 }
 
-FrankLloydRoot.childRoutes = () => ([
+FrankLloydRoot.childRoutes = () => [
   <ModuleRoute path="vitruvius" moduleName="vitruvius-franklin" />,
   <ModuleRoute path="/success" component={HelloWorldComponent} />,
   <ModuleRoute path="healthy-frank" moduleName="healthy-frank" />,
   <ModuleRoute path="demo/:moduleName" moduleName="preview-frank" />,
-]);
+];
 
 FrankLloydRoot.propTypes = {
   children: PropTypes.node.isRequired,