+++ src/server/middleware/sendHtml.js
@@ -16,8 +16,6 @@
 
 // This rule is only needed for a couple functions below
 /* eslint-disable es/no-arrow-functions */
-import { matchesUA } from 'browserslist-useragent';
-import { browserList } from 'babel-preset-amex/browserlist';
 import { Set as iSet, Map as iMap } from 'immutable';
 
 import transit from '../../universal/utils/transit';
@@ -33,6 +31,14 @@ const { buildVersion } = readJsonFile('../../../.build-meta.json');
 const integrityManifest = readJsonFile('../../../bundle.integrity.manifest.json');
 const nodeEnvIsDevelopment = process.env.NODE_ENV === 'development';
 
+// http://www.useragentstring.com/pages/useragentstring.php?name=Internet+Explorer
+const legacyIndicators = [
+  'rv:11', // IE 11 on mobile
+  'MSIE', // IE
+];
+
+const legacyUserAgent = (userAgent) => legacyIndicators
+  .some((legacyIndicator) => userAgent.includes(legacyIndicator));
 function getChunkAssets(assetsByChunkName) {
   return Object
     .entries(assetsByChunkName)
@@ -188,6 +194,7 @@ export function getHead({
   helmetInfo,
   store,
   disableStyles,
+  webManifestUrl,
 }) {
   return `
     <head>
@@ -195,6 +202,7 @@ export function getHead({
       ${disableStyles ? '' : `
       ${renderModuleStyles(store)}
       `}
+      ${webManifestUrl ? `<link rel="manifest" href="${webManifestUrl}">` : ''}
     </head>
   `;
 }
@@ -276,10 +284,7 @@ export default function sendHtml(req, res) {
     } = req;
     const { scriptNonce } = res;
     const userAgent = headers['user-agent'];
-    const isLegacy = !matchesUA(userAgent, {
-      browsers: browserList,
-      allowHigherVersions: true,
-    });
+    const isLegacy = legacyUserAgent(userAgent);
 
     console.info(`sendHtml, have store? ${!!store}, have appHtml? ${!!appHtml}`);
     if (appHtml && typeof appHtml !== 'string') {
@@ -287,7 +292,7 @@ export default function sendHtml(req, res) {
     }
     // replace server specific config with client specific config (api urls and such)
     const clientConfig = getClientStateConfig();
-    const pwaMetadata = getClientPWAConfig();
+    const { webManifestUrl, ...pwaMetadata } = getClientPWAConfig();
     store.dispatch(setConfig(clientConfig));
     const cdnUrl = clientConfig.cdnUrl || '/_/static/';
     const clientInitialState = store.getState();
@@ -313,6 +318,7 @@ export default function sendHtml(req, res) {
       disableScripts,
       disableStyles,
       scriptNonce,
+      webManifestUrl,
     };
 
     const bodySectionArgs = {